Add a build tag to disable all services except TCP endpoint
